requirements.txt
# Stable Diffusion base dependencies
torch==2.6.0
transformers==4.47.1
Pillow==11.0.0
tqdm==4.67.1

# Required for offline tokenizer and checkpoint loading
safetensors>=0.4.1
huggingface-hub>=0.20.0
tokenizers>=0.14.0
